fix for ConnectionData handling
g_dbus_remove_watch() decremented the slot ref count and freed
the ConnectionData associated with it if the ref count reached
zero. This left a dangling pointer to the ConnectionData instance
in the slot, which was found and used once the slot got recycled.

The new approach is to clear the data pointer in the slot if
its last user (watch or handler) is gone.

libgdbus commit ID:
3bb0e1e01230f50f66b3004ca385d73c49bff2c6