--enable-gui=all: use different name in .desktop file for GTK version
When installing both the plain GTK and the Moblin-themed version,
the Moblin version uses the normal "Sync - Up to date" name/comment
and the GTK version uses "Sync (GTK)" as name with the same
comment. This is a somewhat arbitrary choice, with the rationale
being that a Moblin user is less likely to care about the
distinction while a GTK user might understand what "(GTK)" means.