syncevo-dbus-server: implemented syncs via Connection API
At the moment, only running as SyncML client with some
HTTP SyncML server works. It can be triggered by sending
a special message with type "HTTP Config" and the configuration
name as content. dbus-server-http.py can be used to test
this.

The communication between engine and Connection is established
by the new DBusTransportAgent. Except for the limitation
to HTTP client sessions, the Connection API should now
be implemented completely.