command line: add keyring support (MB#3604)
--keyring|-k
  Save or retrieve passwords from the GNOME keyring when modifying the
  configuration or running a synchronization. Note that using this option
  applies to *all* passwords in a configuration, so setting a single
  password as follows moves the other passwords into the keyring, if
  they were not stored there already:
     --keyring --configure --sync-property proxyPassword=foo

  When passwords were stored in the keyring, their value is set to "-"
  in the configuration. This means that when running a synchronization
  without the --keyring argument, the password has to be entered
  interactively.

The implementation introduces new virtual methods for password handling
to properties and iterates over all properties to activate that
special behavior.

Another change is no longer to use cached strings in the SyncEvolutionConfig
to store retrieved passwords. Instead, they are saved as the filter in
the config node. This could help their users to get them but not flush
to files. The main purpose of this change is to make it flexible
to meet the requirements of dynamically added passwords in the backend.