syncevo-dbus-server: allow SAN messages to start a Connection
Server Alerted Notication (SAN) messages (type
"application/vnd.syncml.ds.notification") now are accepted
in the setup phase of a Connection.

The *content* of the message is completely ignored at the
moment. A local config called "default" must exist with
valid settings for each data source, which is used to
to run the sync.