sync progress: generate information about inactive sources
Sources which aren't active during a sync exist in the SyncEvolution
and therefore GUIs may show them. The Synthesis engine isn't told
about these sources, so it won't generate any events for them.

This patch generates the normal 100% done events and the final
source is "done" events inside the EvolutionSyncClient before even
invoking the Synthesis engine. That way GUIs don't have to treat
these inactive sources in any special way.

The command line output is aware of this change and suppresses
output for inactive sources, except for a single INFO line which
tells the user that the source is inactive.

Because of the way the internal API works, it is necessary to
instantiate a dummy SyncSource with no config nodes.