Google template: prevent enabling unsupported sources in GUI by setting empty URIs (MB#5871)
The sync-ui expects that sources which are not currently supported
by a server at all have an empty URI property. This is a reasonable
approach (for example, users can later add a URI and enable the
source, should the server change). However, the Google template
didn't actually set empty URIs for the unsupported calendar,
task and memo sources. Now it does.