D-Bus API: added database listing and config checking
The new Session.GetDatabase() method can be used by a client
to get a list of available local databases, similar to the
output of "syncevolution" without parameters.

Then once a configuration is set (even if only temporarily),
Session.CheckSource() will verify non-destructively whether
the config has a chance of working in a real sync.