SyncMode: use values that are consistent with SyncML spec
This change makes it easier to work with mode numbers coming
from SyncML, like the ones in the Synthesis Server Alerted
Notification (SAN) class.

Previously it was possible to use the enum values as array indices
(SYNC_MODE_MAX). This wasn't used, so it was possible to use larger
values. Replaced SYNC_MODE_MAX with SYNC_INVALID.