syncevo-dbus-server: use Server Alerted Notification (SAN) message content
Selecting a sync mode for all configured sources and selecting
specific sources with one sync mode per source are both supported.

The code is very tolerant of message encoding problems. It logs
errors, then tries to run a sync session with reasonable defaults
despite such errors:
- using a configuration named 'default'
- sync modes as specified in that config

Authorization is not checked at all. In other words, the SAN digest
is ignored. The requested UI mode is also ignored. A sync always runs
in the background, unless a GUI is already running an decides to
display the running sync. Both is consistent with the current design
of SyncEvolution.

The Session ID sent by the server is also ignored. This is an
implementation limitiation which should be fixed. Depends on
a suitable API for the Synthesis Engine.

The Server ID is used to find a corresponding local configuration
as follows:
- first match against syncURL of all configured servers
- if not found, match against the config names themselves

The local configuration must exist. It is not currently created
on-the-fly.

If the SAN message specifies specific URIs, then the corresponding
local data sources are identified by comparing the server's URI
against the beginning of the URI in the local source
configuration. That way a local URI may contain parameters appended to
it (supported by some servers to select the data which is to be
synchronzied).