message resend: must set transport error when giving up
The message resend code aborted the sync by asking the engine
to abort via the step command. When the initial message send
failed, the engine didn't generate an error code and thus the
overall sync result was "okay". It should have been "transport
failed".

This patch ensures that the transport failure is recorded by
throwing a TransportError. This is caught, sets the sync status,
then aborts with the step command as before.