config: allow empty username/password
Leaving the username/password empty might be valid for servers.
With the previous non-empty defaults for these two properties it
was not possible to set empty values, because setting empty values
was treated like choosing the default.

As a fix the default is now empty. The previous default values
were moved into the server templates.