RetryInterval/Delay: only accept unsigned integers (MB #6500)
Negative values silently disabled message resending. This was
not documented and thus unexpected, so trigger an error for
this kind of values now (using the new type checking for
integer values).

To disable the feature users would have to use very large
integer numbers (not documented, but kind of obvious I hope).