--enable-evolution-compatibility: added icalcomp_get_location/summary (MB#6552), added test
Two more functions which have to be intercepted to avoid hard
dependencies on libecal or libical.

Because this is something which is likely to come up again
in the future, this patch also adds a "installcheck-local" for
undesired dependencies on libical/ecal/ebook/edataserver.

This check is triggered when building binary packages by using
"installcheck" instead of "install". It intentionally checks
.so files in the build directory, to catch problems even in
libs which are not currently installed.