packaging: fix for installcheck
First, installcheck does not imply install. Packaging must
ask for both.

Second, installcheck of header files and libs must be aware
of DESTDIR. src/syncevo/installcheck-local.sh patches the -I and -L
parameters so that it includes the DESTDIR passed to it.

Third, "distbin" overwrites "prefix", which breaks this. Don't
try installcheck as part of that target.