syncevo-dbus-server: cleaned up read-only methods and added GetConfigs()
With the new binding support in the D-Bus C++ bridge the glue code
in Session became obsolete. There's also no reason anymore to
derive Session from ReadOperations. Access to the config name
used by the session is now done via getConfig(), which hides
the implementation details.

The GetConfigs() call defined in the API was missing in the
implementation. Added as member of the ReadOperations,
making it static because it doesn't need a config name.