packaging: another fix for installcheck-local
When the installation is in the standard location, pkg-config won't
print -I and -L options. The "redirect into DESTDIR via sed" trick
depended on that, meaning that the files were not found during the
nightly build.

Now -I and -L for the DESTDIR dirs is added explicitly. LD_LIBRARY_PATH
also has to be set, for libsmltk to be found.