Mobical: PHOTOs preserved verbatim (MB#6668)
Originally, testing against Mobical led to the conclusion that
   'PHOTO' is compressed by server but still be correct
and checking of the PHOTO property was disabled in synccompare.

Later it was found that there were decoding (Mobical
not standard compliant) and encoding (libsynthesis not
compliant when folding was off) issues (MB#6668), fixed in
libsynthesis now.

Automated testing no longer falls over a modified PHOTO property.
Either the compression is not triggered for this photo, or
something has changed. In any case, this patch enables comparison
of the PHOTO property again, because that is necessary to cover
these encoding issues.