client-test: added testManyDeletes, improved testManyItems
testManyItems now checks that the statistics are correct. This
is based on the assumption that all active sources use the same
number of test items.

The new testManyDeletes triggers a problem with 222 loop detection:
when the message to the server contains many deletes, the server
may require so many reply messages, that the client has to send
5 Alert 222 messages in a row and incorrectly treats this as a loop
that has to be aborted.