syncevo-dbus-server: Server.GetConfig() should not create configs
The Server.GetConfig() is stricly read-only. When looking at the
error message for NoSuchConfig I noticed that it creates the config
from a template with the name if the config doesn't exist yet. This
was discussed before and should have been removed, but this instance
must have slipped in.

Removed and adapted test for error message accordingly.