cmdline --print-servers: sort by server name
The order of the servers dependend on the order in which the
filesystem returned the directory names, in other words, it
was pretty random. Now the list is sorted alphabetically by
the server name.

This was found in the testPrintServers when the order of
the servers started to deviate from the (also unsorted) order
in the reference output.