Server Alerted Sync: SAN generation
Use "peerIsClient" configuration property to decide whether this should be a
Server Alerted Sync Session or a client initated.

Use "remoteIdentifier" as the string sending to the remote peer during SAN so
that remote peer can identifier the server properly.

If this should be, will first generate and send SAN in SyncContext::doSync,
get the client request and inialize the session as server with the received
request.